+++ R/screenshot.R
@@ -49,7 +49,7 @@ screenshot <- function(file = NULL, proj = proj_get(), dir = NULL) {
   proj_path <- proj_list(proj)
 
 
-  if (!rstudioapi::isAvailable()) {
+  if (!is_rstudio()) {
     cli::cli_warn("This feature may not work as excepted outside RStudio.")
   }
 
@@ -196,3 +196,4 @@ screenshot <- function(file = NULL, proj = proj_get(), dir = NULL) {
   cli::cli_inform(bullets)
   invisible(img_path_rel_chr)
 }
+